feat(mito2): align nested JSON2 parquet schemas
